docker_provider/requirements.txt
websocket-client==1.2.1
PyYAML==5.3.1
pymongo==3.12.0
simplejson==3.16.0
pandas==1.3.3
numpy==1.21.2
openpyxl==3.0.9
python-binance==1.0.15
Flask==2.0.2